requirements.txt
numpy>=1.26.0
pandas>=2.0.0
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
tqdm>=4.66.1
requests>=2.31.0
snowflake-connector-python>=3.0.0
pyarrow>=14.0.1